requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coverage==7.6.12
defusedxml==0.7.1
Django==5.1.6
django-modeltranslation==0.19.12
django-parler==2.3
django-restframework==0.0.1
django-rosetta==0.10.1
djangorestframework==3.15.2
httpie==3.2.4
idna==3.10
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
polib==1.2.0
Pygments==2.19.1
PySocks==1.7.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0

drf-yasg~=1.21.8